requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
boto3==1.40.11
google-api-python-client==2.176.0
google-auth==2.40.3
python-multipart==0.0.20
